requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
groq==0.13.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-resources==6.4.5
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
narwhals==1.16.0
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil-resolve-name==1.3.10
propcache==0.2.0
protobuf==5.29.1
pyarrow==17.0.0
pydantic==2.10.3
pydantic-core==2.27.1
pydeck==0.9.1
pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.15.2
zipp==3.20.2
pydantic-settings